refactor(stripe-issuing): split orchestration class to gateway, keep primitives in SDK
Boundary-audit follow-up. The stripe-issuing module was the only file
flagged AMBIGUOUS in the 60-file boundary audit because it mixed
protocol primitives (signature verifier, V2Delegation projection,
spending-controls mapper, form encoder, constants) with product
runtime (live api.stripe.com HTTP client, sk_test_ credential
handling, in-memory card↔delegation registry, the orchestration
class). LF takes the SDK on AAIF acceptance; we do not transfer
credential-handling code.

SDK-side changes:

  - New `src/v2/payment-rails/stripe-issuing/primitives.ts` carrying
    `defaultMapDelegationToSpendingControls`, `delegationToView` (now
    exported), `encodeForm` (now exported), `verifyStripeSignature`,
    and the four constants (DEFAULT_API_BASE, DEFAULT_REQUIRED_SCOPE,
    DEFAULT_TOLERANCE_SEC, RAIL_NAME).

  - `src/v2/payment-rails/stripe-issuing/index.ts` is now a thin
    barrel that re-exports primitives.ts. Existing import paths
    through this file for the primitive surface continue to work.
    The class `StripeIssuingRail` and its `createStripeIssuingRail`
    factory are NO LONGER exported here — they have moved to the
    private gateway repo.

  - `src/v2/payment-rails/index.ts` and `src/index.ts` drop the class
    exports. New aliased exports added so the gateway can pull
    `delegationToView` and `encodeForm` from the npm package surface
    without name collisions: `stripeIssuingDelegationToView`,
    `stripeIssuingEncodeForm`, plus the four
    STRIPE_ISSUING_DEFAULT_* constants.

  - Tests: the class-exercising tests (provisionAgentCard,
    handleAuthorizationWebhook approve/decline/escalation paths,
    revokeWallet, verifyTransaction, fixture exercise, PaymentRail
    surface, sk_test_ live smoke) moved out of the SDK suite. The
    primitive tests (mapper, standalone signature verifier) live in
    a new `stripe-issuing-primitives.test.ts`. Class tests preserved
    at /tmp for placement in the gateway repo on a separate branch.

Bytewise output of the protocol-defining functions
(`defaultMapDelegationToSpendingControls`, `verifyStripeSignature`)
is unchanged. Conformance tests pin them.

Test count: 2,884 → 2,856 (-28). Net: 36 class tests moved out, 8
primitive tests added (4 mapper + 4 signature standalone). Suite
pass: 2,856/2,856, 0 fail, 0 skipped.
